Add files via upload
Collect keystrokes from a 4 x 4 matrix keypad, and use them to control number of wires and length in inches of the wire to be made... on a Wire Making Machine.
This portion of the code shows what I did to get usable data.  variable "wires" and "inches' are what is used for control elsewhere in the program, not shown here.